Move PointOfSaleSettings class into own file
